+++ CHANGELOG.md
@@ -14,8 +14,8 @@ Unreleased
   - Removed the display of default profile.
 - Improved the display of help messages.
   - Added global option separated from options.
-  - Changed the color of option metadata.
   - Changed max width to 100 to display to reduce line breaks.
+  - Removed the color of option type.
 - Removed the unavailable options.
 - Fixed the required Python version to 3.6.1 or higher.